+++ app/services/player_service.py
@@ -198,7 +198,7 @@ def get_fishing_preparation(user_id):
     #更新GMC
     #从合约更新user_gmc
     gmc_contract = ethereum_service.get_gmc_contract()
-    user.user_gmc = gmc_contract.functions.balanceOf(user_id).call() * (10 ** -18)  # .call() 用于在本地执行合约函数，不会发起链上交易
+    user.user_gmc = int(gmc_contract.functions.balanceOf(user_id).call() * (10 ** -18))  # .call() 用于在本地执行合约函数，不会发起链上交易
     # #从合约更新鱼饵数量
     # user_contract = ethereum_service.get_user_contract()
     # user.user_baits = user_contract.functions.getBaitCount(user_id).call()
@@ -252,8 +252,8 @@ def get_fishing_preparation(user_id):
             'fishs_max': current_pond.fishs_max,
             'user_exp': user.user_exp,
             'max_exp': level_exp.max_exp,
-            'user_gmc': float(user.user_gmc),
-            'collected_gmc': float(user.collected_gmc),
+            'user_gmc': user.user_gmc,
+            'collected_gmc': user.collected_gmc,
             'bubble_gmc': user.bubble_gmc,
             'bubble_gmc_max': bubble_gmc_max,
             'user_baits': user.user_baits,
